remove old `Score` implementation and add some regression tests to search.rs
